+++ src/css/styles.scss
@@ -13,6 +13,7 @@ $base-text-size: 18px;
   --il-content-max-width: 1140px;
   --il-focus-outline: 3px dashed var(--il-focus-outline-color);
   --il-dotted-focus-outline: 4px dotted var(--il-focus-outline-color);
+  --il-first-child-margin-top: 0;
 
   @media (min-width: 768px) {
     --il-content-margin: 30px;
