+++ src/css/typography/_index.scss
@@ -7,7 +7,7 @@
   padding: 0;
 
   &:first-child {
-    margin-top: 0;
+    margin-top: var(--il-first-child-margin-top);
   }
 }
