+++ src/js/components/video-feature/video-feature.component.js
@@ -55,23 +55,22 @@ class VideoFeatureComponent extends LitElement {
     }
     var contentClass = (this.background == 'solid' || this.background == 'blue') ? 'content solid' : this.background == 'white' ? 'content' : this.background == 'orange' ? 'content orange' : this.background == 'gray' ? 'content white gray' :'content gradient';
     var backgroundClass = (this.background == 'solid' || this.background == 'blue') ? 'background solid' : this.background == 'white' ? 'background' : this.background == 'orange' ? 'background orange' : this.background == 'gray' ? 'background gray' : 'background gradient';
-    var leftFlex = this.size == 'large' ? 2 : 1;
-    var rightFlex = this.size == 'small' ? 2 : 1;
+
     if (this.align == 'right')
     {
         return html`
         <div class="videofeature">
-            <div class="${contentClass}" style="flex: ${rightFlex};">
+            <div class="${contentClass}">
                 <slot></slot>
             </div>
-            <div class=${backgroundClass}" style="flex: ${leftFlex}; display: flex;">${this.renderVideo()}</div>
+            <div class=${backgroundClass}">${this.renderVideo()}</div>
         </div>
         `;
     }
     return html`
         <div class="videofeature">
-            <div class="${backgroundClass}" style="flex: ${leftFlex}; display: flex;">${this.renderVideo()}</div>
-            <div class="${contentClass}" style="flex: ${rightFlex};">
+            <div class="${backgroundClass}">${this.renderVideo()}</div>
+            <div class="${contentClass}">
                 <slot></slot>
             </div>
         </div>
