+++ src/js/index.js
@@ -5,11 +5,11 @@ import './components/back-to-top';
 import './components/breadcrumbs-page';
 import './components/breadcrumbs';
 import './components/call-to-action';
-import './components/clickable-card';
+import './components/clickable-card/clickable-card.component';
 import './components/footer';
 import './components/header';
 import './components/hero';
-import './components/image-feature';
+import './components/image-feature/image-feature.component';
 import './components/layout/page.component';
 import './components/layout/section.component';
 import './components/layout/section-with-sidebar.component';
