+++ tests/components/clickable-card/attributes.njk
@@ -1,5 +1,5 @@
 ---
-title: Clickable Card (default view)
+title: Clickable Card (depreciated attributes)
 group: "Components: Clickable Card"
 layout: layouts/page.njk
 ---
@@ -13,7 +13,7 @@ the classroom through internships, judging teams, student organizations, undergr
 projects with our faculty, as well as short- and long-term study abroad opportunities all over the world. </p>
 </il-clickable-card>
 
-<il-clickable-card width="500px" src="{{ '/_assets/images/hero-bg.jpg' | url }}" href="https://www.example.com" class="il-theme-blue">
+<il-clickable-card width="500px" src="{{ '/_assets/images/hero-bg.jpg' | url }}" href="https://www.example.com" background="blue">
 <h3 slot="header">Student Life</h3>
 <p>Animal sciences students extend their learning and career networks beyond
 the classroom through internships, judging teams, student organizations, undergraduate research
