+++ tests/components/clickable-card/imagefilter.njk
@@ -18,7 +18,7 @@ the classroom through internships, judging teams, student organizations, undergr
 projects with our faculty, as well as short- and long-term study abroad opportunities all over the world. </p>
 </il-clickable-card>
 
-<il-clickable-card width="500px" src="{{ '/_assets/images/hero-bg.jpg' | url }}" href="https://www.example.com" background="blue">
+<il-clickable-card width="500px" src="{{ '/_assets/images/hero-bg.jpg' | url }}" href="https://www.example.com" class="il-theme-blue">
 <h3 slot="header">Student Life</h3>
 <p>Animal sciences students extend their learning and career networks beyond
 the classroom through internships, judging teams, student organizations, undergraduate research
