+++ tests/components/clickable-card/nopicture.njk
@@ -1,0 +1,16 @@
+---
+title: Clickable Card (no picture)
+group: "Components: Clickable Card"
+layout: layouts/page.njk
+---
+
+<div class="il-formatted">
+
+<il-clickable-card width="500px" href="https://www.example.com">
+<h3 slot="header">Student Life</h3>
+<p>Animal sciences students extend their learning and career networks beyond
+the classroom through internships, judging teams, student organizations, undergraduate research
+projects with our faculty, as well as short- and long-term study abroad opportunities all over the world. </p>
+</il-clickable-card>
+
+</div>
