+++ tests/components/image-feature/attributes.njk
@@ -1,87 +1,95 @@
 ---
-title: Image Feature (default view)
+title: Image Feature (depreciated attributes)
 group: "Components: Image Feature"
 layout: layouts/page.njk
 ---
-<h2>Gradient</h2>
 <div class="il-formatted">
-<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" alt="Adding accessibility information describing the image here">
+<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" align="left" alt="Adding accessibility information describing the image here">
 <h3>Student Life</h3>
 <p>Animal sciences students extend their learning and career networks beyond
 the classroom through internships, <a href="#">judging teams</a>, student organizations, undergraduate research
 projects with our faculty, as well as short- and long-term study abroad opportunities all over the world. </p>
+
 <a href="#" class="il-button">Subscribe</a>
 <a href="#" class="il-button">Find a departmental advisor</a>
+ 
 </il-image-feature>
 
-<h2>Align Right</h2>
-<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" class="il-align-right">
+<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" align="right">
 <h3>Student Life</h3>
 <p>Animal sciences students extend their learning and career networks beyond
 the classroom through internships, <a href="#">judging teams</a>, student organizations, undergraduate research
 projects with our faculty, as well as short- and long-term study abroad opportunities all over the world. </p>
+
 <a href="#" class="il-button">Subscribe</a>
 <a href="#" class="il-button">Find a departmental advisor</a>
+ 
 </il-image-feature>
 
-<h2>White</h2>
-<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" class="il-theme-white">
+<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" background="white" align="left">
 <h3>Student Life</h3>
 <p>Animal sciences students extend their learning and career networks beyond
 the classroom through internships, <a href="#">judging teams</a>, student organizations, undergraduate research
 projects with our faculty, as well as short- and long-term study abroad opportunities all over the world. </p>
+
 <a href="#" class="il-button">Subscribe</a>
 <a href="#" class="il-button">Find a departmental advisor</a>
+ 
 </il-image-feature>
 
-<h2>Small</h2>
-<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" class="il-size-small">
+<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" size="small" align="left">
 <h3>Student Life</h3>
 <p>Animal sciences students extend their learning and career networks beyond
 the classroom through internships, <a href="#">judging teams</a>, student organizations, undergraduate research
 projects with our faculty, as well as short- and long-term study abroad opportunities all over the world. </p>
+
 <a href="#" class="il-button">Subscribe</a>
 <a href="#" class="il-button">Find a departmental advisor</a>
+ 
 </il-image-feature>
 
-<h2>Large, Blue</h2>
-<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" class="il-size-large il-theme-blue">
+<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" size="large" align="left" background="solid">
 <h3>Student Life</h3>
 <p>Animal sciences students extend their learning and career networks beyond
 the classroom through internships, <a href="#">judging teams</a>, student organizations, undergraduate research
 projects with our faculty, as well as short- and long-term study abroad opportunities all over the world. </p>
+
 <a href="#" class="il-button">Subscribe</a>
 <a href="#" class="il-button">Find a departmental advisor</a>
+ 
 </il-image-feature>
 
-<h2>Extra small, white</h2>
-<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" class="il-size-xsmall il-theme-white">
+<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" size="portrait" align="left" background="white">
 <h3>Student Life</h3>
 <p>Animal sciences students extend their learning and career networks beyond
 the classroom through internships, <a href="#">judging teams</a>, student organizations, undergraduate research
 projects with our faculty, as well as short- and long-term study abroad opportunities all over the world. </p>
+
 <a href="#" class="il-button">Subscribe</a>
 <a href="#" class="il-button">Find a departmental advisor</a>
+ 
 </il-image-feature>
 
-<h2>Large, right</h2>
-<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" class="il-size-large il-align-right">
+<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" size="large" align="right">
 <h3>Student Life</h3>
 <p>Animal sciences students extend their learning and career networks beyond
 the classroom through internships, <a href="#">judging teams</a>, student organizations, undergraduate research
 projects with our faculty, as well as short- and long-term study abroad opportunities all over the world. </p>
+
 <a href="#" class="il-button">Subscribe</a>
 <a href="#" class="il-button">Find a departmental advisor</a>
+ 
 </il-image-feature>
 
-<h2>Orange</h2>
-<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" class="il-theme-orange">
+<il-image-feature src="{{ '/_assets/images/hero-bg.jpg' | url }}" background="orange" align="left">
 <h3>Student Life</h3>
 <p>Animal sciences students extend their learning and career networks beyond
 the classroom through internships, <a href="#">judging teams</a>, student organizations, undergraduate research
 projects with our faculty, as well as short- and long-term study abroad opportunities all over the world. </p>
+
 <a href="#" class="il-button">Subscribe</a>
 <a href="#" class="il-button">Find a departmental advisor</a>
+ 
 </il-image-feature>
 
 </div>
