+++ tests/components/video-feature/attributes.njk
@@ -1,12 +1,12 @@
 ---
-title: Video Feature (default view)
+title: Video Feature (using depreciated attributes)
 group: "Components: Video Feature"
 layout: layouts/page.njk
 ---
 <div class="il-formatted">
 <h2>Gradient</h2>
 <p><em>Pulled out video for visual testing -- can use https://mediaspace.illinois.edu/media/t/1_0o2ypqvp</em></p>
-<il-video-feature src="">
+<il-video-feature src="" align="left" data-il-debug="true">
 <h3>Student Life</h3>
 <p>Animal sciences students extend their learning and career networks beyond
 the classroom through internships, judging teams, student organizations, undergraduate research
@@ -17,7 +17,7 @@ projects with our faculty, as well as short- and long-term study abroad opportun
 
 <h2>Orange, Right</h2>
 <p><em>Pulled out video for visual testing -- can use https://www.youtube.com/watch?v=uSEpvhL7nwA</em></p>
-<il-video-feature src="" class="il-align-right il-theme-orange">
+<il-video-feature src="" align="right" background='orange' data-il-debug="true">
 <h3>Student Life</h3>
 <p>Animal sciences students extend their learning and career networks beyond
 the classroom through internships, judging teams, student organizations, undergraduate research
@@ -28,7 +28,7 @@ projects with our faculty, as well as short- and long-term study abroad opportun
 
 <h2>White</h2>
 <p><em>Pulled out video for visual testing -- can use https://mediaspace.illinois.edu/media/t/1_lopu8gqs</em></p>
-<il-video-feature src="" class="il-align-right il-theme-white" view='vertical'>
+<il-video-feature src="" align="left" background='white' view='vertical' data-il-debug="true">
 <h3>Student Life</h3>
 <p>Animal sciences students extend their learning and career networks beyond
 the classroom through internships, judging teams, student organizations, undergraduate research
@@ -39,7 +39,7 @@ projects with our faculty, as well as short- and long-term study abroad opportun
 
 <h2>Vimeo</h2>
 <p><em>Pulled out video for visual testing -- can use https://vimeo.com/17327293</em></p>
-<il-video-feature src="" class="il-align-left il-theme-white" view='vertical'>
+<il-video-feature src="" align="left" background='white' view='vertical' data-il-debug="true">
 <h3>Student Life</h3>
 <p>Animal sciences students extend their learning and career networks beyond
 the classroom through internships, judging teams, student organizations, undergraduate research
