+++ tests/components/video-feature/attributes.vis.js
@@ -1,0 +1,12 @@
+const util = require('../../tests.util');
+
+const url = util.testUrl(__filename);
+
+module.exports = (viewports) => {
+  return [
+    {
+      url, label: "videofeature-attributes",
+      viewports: [viewports.desktop, viewports.iphone, viewports.hdtv]
+    }
+  ]
+}
